op-e2e: Move super root helper out of op-service into op-e2e
